Adding support for S3 deployment, along with some other cleanup.
