+++ package.json
@@ -2,7 +2,7 @@
   "name": "constitutional",
   "version": "1.0.0",
   "description": "Constitution of the United States of America",
-  "main": "index.js",
+  "main": "client.js",
   "directories": {
     "test": "test"
   },
@@ -12,7 +12,7 @@
   "author": "Trevor O'Brien and Luke Bayes",
   "license": "MIT",
   "dependencies": {
-    "nomplate": "^1.1.3"
+    "nomplate": "^1.1.13"
   },
   "devDependencies": {
     "babel-core": "^6.20.0",
