--- static/index.html
@@ -1,11 +1,0 @@
-<html>
-  <!-- TODO(lbayes): Create nomplate/cli.js to accept an input file and emit this static document -->
-  <head>
-    <title>Constitution of the United State</title>
-    <link rel="stylesheet" href="constitutional.css">
-    <script src="constitutional.js"></script>
-  </head>
-  <body>
-    <div id="container"></div>
-  </body>
-</html>
